Moved retry classes into a retry module
