+++ symphony-bdk-core/src/main/java/com/symphony/bdk/core/service/SessionService.java
@@ -1,10 +1,8 @@
 package com.symphony.bdk.core.service;
 
-import com.symphony.bdk.core.api.invoker.ApiException;
-import com.symphony.bdk.core.api.invoker.ApiRuntimeException;
 import com.symphony.bdk.core.auth.AuthSession;
 import com.symphony.bdk.core.config.model.BdkRetryConfig;
-import com.symphony.bdk.core.util.function.RetryWithRecovery;
+import com.symphony.bdk.core.retry.RetryWithRecovery;
 import com.symphony.bdk.gen.api.SessionApi;
 import com.symphony.bdk.gen.api.model.UserV2;
